requirements.txt
torch
torchvision
torchaudio
diffusers
transformers
accelerate
safetensors
Pillow
streamlit
python-dotenv
tqdm
